// gui/corpus-gui/src/components/ScatterPlot/SelectionPlugin/types.ts
import type { UmapPoint } from "../types";

export type Id = UmapPoint["eventId"];

export interface ScreenRect {
	x: number;
	y: number;
	width: number;
	height: number;
}

export interface SelectionEvent<T> {
	type: "click" | "rect" | "background-click" | "null-select";
	payload: T;
}

export type SelectionMode = "replace" | "additive";

export interface SelectionOptions {
	mode?: SelectionMode;
	multiKey?: "Shift" | "Ctrl" | "Alt";
}

export interface SelectionPlugin {
	destroy(): void;
}